requirements.txt
pandas
requests
numpy
matplotlib
seaborn
scikit-learn==1.6.0
flask
fastapi
uvicorn
ipykernel
catboost
cassandra-driver
xgboost
plotly
streamlit
folium
streamlit_folium